Refactored commong behavior to common test module
