+++ engine/backend/common/src/test/java/com/torodb/backend/common/AbstractMetaDataIT.java
@@ -1,22 +1,4 @@
-/*
- * ToroDB
- * Copyright © 2014 8Kdata Technology (www.8kdata.com)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.torodb.backend.derby;
+package com.torodb.backend.common;
 
 import com.torodb.backend.SqlInterface;
 import com.torodb.backend.tables.MetaCollectionTable;
@@ -37,23 +19,27 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class MetaDataIT {
+public abstract class AbstractMetaDataIT {
 
   private SqlInterface sqlInterface;
 
-  private DatabaseTextContext dbTestContext;
+  private DatabaseTestContext dbTestContext;
 
   @Before
   public void setUp() throws Exception {
-    dbTestContext = new DatabaseTestContextFactory().createDerbyInstance();
+    dbTestContext = getDatabaseTestContext();// new DatabaseTestContextFactory().createDerbyInstance();
     sqlInterface = dbTestContext.getSqlInterface();
     dbTestContext.setupDatabase();
   }
+
+  protected abstract DatabaseTestContext getDatabaseTestContext();
 
   @After
   public void tearDown() throws Exception {
     dbTestContext.tearDownDatabase();
   }
+
+
 
   @Test
   public void metadataDatabaseTableCanBeWritten() throws Exception {
