+++ engine/backend/derby/src/test-integration/java/com/torodb/backend/derby/DerbyMetaDataIT.java
@@ -1,0 +1,32 @@
+/*
+ * ToroDB
+ * Copyright © 2014 8Kdata Technology (www.8kdata.com)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.torodb.backend.derby;
+
+import com.torodb.backend.common.AbstractMetaDataIT;
+import com.torodb.backend.common.DatabaseTestContext;
+
+public class DerbyMetaDataIT extends AbstractMetaDataIT {
+
+  @Override
+  protected DatabaseTestContext getDatabaseTestContext() {
+    return new DerbyDatabaseTestContextFactory().createInstance();
+//    return new DatabaseTestContextFactory().createDerbyInstance();
+  }
+
+}
